add jbuilder to return comapny json
