+++ app/views/api/companies/_company.json.jbuilder
@@ -1,0 +1,11 @@
+json.call(
+  company,
+  :id,
+  :ruc,
+  :economic_activity,
+  :contributor_type,
+  :legal_representative,
+  :business_name,
+  :logo,
+  :slogan
+)
