+++ app/views/api/companies/create.json.jbuilder
@@ -1,0 +1,6 @@
+json.company do
+  json.partial!(
+    'api/companies/company',
+    company: company
+  )
+end
