+++ config/routes.rb
@@ -13,9 +13,7 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
   namespace :api, defaults: { format: :json } do
-    namespace :user do
-      resources :companies,
-                only: [:create, :update, :show]
-    end
+    resources :companies,
+              only: [:create, :update, :show]
   end
 end
